Split code into separate files due to GodClass violation
