+++ src/test/java/no/sikt/nva/validators/ExcelTestingUtils.java
@@ -1,0 +1,74 @@
+package no.sikt.nva.validators;
+
+import java.time.Instant;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Date;
+import java.util.List;
+import no.sikt.nva.brage.migration.common.model.record.BrageVersion;
+import no.sikt.nva.brage.migration.common.model.record.license.BrageLicense;
+import org.apache.poi.ss.usermodel.Cell;
+import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.ss.usermodel.Workbook;
+import org.apache.poi.xssf.usermodel.XSSFWorkbook;
+
+public class ExcelTestingUtils {
+    public static final Object[] VALID_ROW = {
+        123,
+        "En tittel",
+        BrageVersion.ACCEPTED.getValue(),
+        Date.from(Instant.now()),
+        BrageLicense.CC_BY.getValue(),
+        "Fil.pdf"
+    };
+    public static final String INVALID_DATE_FORMAT_EXAMPLE = "03/19/2024";
+    public static final String EMPTY_STRING = "";
+
+    public static Workbook createWorkbook(Object[][] bookData) {
+        var workbook = new XSSFWorkbook();
+        var sheet = workbook.createSheet("Metadata-test");
+
+        var createHelper = workbook.getCreationHelper();
+        var styleDate = workbook.createCellStyle();
+        styleDate.setDataFormat(createHelper.createDataFormat().getFormat("dd/MM/yyyy"));
+
+        int rowCount = 0;
+
+        for (Object[] aBook : bookData) {
+            Row row = sheet.createRow(rowCount++);
+
+            int columnCount = 0;
+
+            for (Object field : aBook) {
+                Cell cell = row.createCell(columnCount++);
+                if (field instanceof String) {
+                    cell.setCellValue((String) field);
+                } else if (field instanceof Integer) {
+                    cell.setCellValue((Integer) field);
+                } else if (field instanceof Date) {
+                    cell.setCellStyle(styleDate);
+                    cell.setCellValue((Date) field);
+                }
+            }
+        }
+
+        return workbook;
+    }
+
+    public static Object[] substituteValue(Object[] list, Object newValue, int index) {
+        list[index] = newValue;
+        return list;
+    }
+
+    public static Object[] addValue(Object[] list, Object newValue) {
+        var arrayWithExtraValue = Arrays.copyOf(list, list.length + 1);
+        arrayWithExtraValue[arrayWithExtraValue.length - 1] = newValue;
+        return arrayWithExtraValue;
+    }
+
+    public static Object[] removeValue(Object[] list, int index) {
+        var newArray = new ArrayList<>(List.of(list));
+        newArray.remove(index);
+        return newArray.toArray();
+    }
+}
